Move module and feature instantiation in deployment scripts
