+++ deployment/3_setup_contracts.js
@@ -1,7 +1,7 @@
 const ModuleRegistry = require("../build/ModuleRegistry");
 const ENSManager = require("../build/ArgentENSManager");
 const ENSResolver = require("../build/ArgentENSResolver");
-const WalletFactory = require("../build-legacy/v1.6.0/WalletFactory");
+const WalletFactory = require("../build/WalletFactory");
 const TokenPriceRegistry = require("../build/TokenPriceRegistry");
 const CompoundRegistry = require("../build/CompoundRegistry");
 const DexRegistry = require("../build/DexRegistry");
