+++ scripts/deploy.sh
@@ -21,10 +21,7 @@ else
     shift
 fi
 
-rm -rf build
-npm run compile:lib
-npm run compile
-npm run compile:legacy
+
 
 for IDX in "$@"
 do
