+++ utils/config/ganache.json
@@ -1,1 +1,1 @@
-{"ENS":{"deployOwnRegistry":true,"ensRegistry":"0x54453dF26C5C638C68dF46e3f1B83360b0Ab29fd","domain":"argent.xyz"},"backend":{"accounts":["0xD9995BAE12FEe327256FFec1e3184d492bD94C31"]},"multisig":{"owners":["0xD9995BAE12FEe327256FFec1e3184d492bD94C31"],"threshold":1,"autosign":true},"settings":{"deployer":{"type":"ganache"},"lockPeriod":480,"recoveryPeriod":480,"securityPeriod":240,"securityWindow":240,"feeRatio":15,"defaultLimit":"1000000000000000000"},"CryptoKitties":{"contract":"0x0000000000000000000000000000000000000000"},"defi":{"weth":"0x0000000000000000000000000000000000000000","maker":{"deployOwn":true,"tub":"0x0000000000000000000000000000000000000000","pot":"0x0000000000000000000000000000000000000000","jug":"0x0000000000000000000000000000000000000000","migration":"0xC07f212AaEeF5Dc948BBC5302cB46D9D799417Ec"},"uniswap":{"deployOwn":true,"factory":"0x4d5440AC7CEA2DDBc92Fe35756c90EEBbAe58cEb"},"compound":{"comptroller":"0x0000000000000000000000000000000000000000","markets":{"0x0000000000000000000000000000000000000000":"0x0000000000000000000000000000000000000000"}},"paraswap":{"deployOwn":true,"contract":"0x41B10475131cA22A3D574DEa3B3670B5ab2AD90E","authorisedExchanges":{"Kyber":"0x8f8f0454f35F5daE79E6129329Dd3C80AEc3A8cc"}}},"contracts":{"MultiSigWallet":"0x019e04b6fc11305c273924B58F93e8c7ecF934B9","WalletFactory":"0x2F2c4e8e7b8c31d836e1158A10902b504D0Bc71F","ENSResolver":"0x061ABd0F2579b1A41a3C8b741a74485eA96e9AF0","ENSManager":"0x39641b8415416d924BC625362571826485e6DB1b","TokenPriceProvider":"0xE09C322B745434Fbb2D18c107c13795D4bDF3657","ModuleRegistry":"0x6B74146d42286a96C656f0368C0dE4A43c3b546a","BaseWallet":"0x406F6ae39429b35853d54e1d75De71F793fB5C6b","CompoundRegistry":"0xfDEd0b9b98a28964208A3c3a1351F7414eD743EC","MakerRegistry":"0x355d26E5f04255F60E910F1aA1251DE5D10996E8","DexRegistry":"0xbf7E0dcC7a3D725Fb0768836cfd32c13a9ef1324"},"modules":{"GuardianStorage":"0x6c4539ea80eFF2b0FA546ca0A795A8Cf84d1b31C","TransferStorage":"0x3Ab452E9436fCc8Ba1CA86F9112Ffe18653Ca6Be","GuardianManager":"0xa375B16db7412f1104A8820f8C7485F19dA5deA0","LockManager":"0x215a1EfB566aA85c756C69Fc748E584AEc13CCb1","RecoveryManager":"0x1E8c7A43E530c73FA69F0DE1805AABa73336eD34","ApprovedTransfer":"0x6CaDE3FAa7FF8CA5961AB26F5f25D56E7420f584","TokenExchanger":"0xC8ECF2fb6DE001995D35FB707E2655B070b7a06E","NftTransfer":"0x33454011c495540CC30bfFaed39644F0F779cf73","MakerManager":"0x2718cAce241213B18F67C6d66B245EE77DB0Ce2a","TransferManager":"0x47C5A7F8F78c531c789A25adACf8169cA1d8B173","CompoundManager":"0xD33fA1776ff3073d110874eBe54f405Ae10BbD46","MakerV2Manager":"0x0F0A56e838323Df990DDD41cd95e4A93942bC6ff","LimitStorage":"0x9A9BA9ec5A949f453162fa6C0541F580162Ae400","TokenPriceStorage":"0x5dA4F81c17776495e3d7999ad9217748D257f074","RelayerModule":"0x171eb25c80da6adF0436668126285ca51e37eB30","RelayerManager":"0xC9737dA2aAFA08F296805C50bE83A7086682beAf","VersionManager":"0x320762f547A460453A1115251F72F6E4daef3c6c","LockStorage":"0x189F52364c5fC5dd16600b42094F2a04078AE928"},"gitCommit":"37220550ce6b4dbe303069d309ea56765e48be3a"}
+{"ENS":{"deployOwnRegistry":true,"ensRegistry":"0x7E93afFaa1452a7e07B961Aab04EF20a4Bc12b78","domain":"argent.xyz"},"backend":{"accounts":["0xD9995BAE12FEe327256FFec1e3184d492bD94C31"]},"multisig":{"owners":["0xD9995BAE12FEe327256FFec1e3184d492bD94C31"],"threshold":1,"autosign":true},"settings":{"deployer":{"type":"ganache"},"lockPeriod":480,"recoveryPeriod":480,"securityPeriod":240,"securityWindow":240,"feeRatio":15,"defaultLimit":"1000000000000000000"},"CryptoKitties":{"contract":"0x0000000000000000000000000000000000000000"},"defi":{"weth":"0x0000000000000000000000000000000000000000","maker":{"deployOwn":true,"tub":"0x0000000000000000000000000000000000000000","pot":"0x0000000000000000000000000000000000000000","jug":"0x0000000000000000000000000000000000000000","migration":"0x101b60626A8073271fE2eFE59FB8fD74c108c2eC"},"uniswap":{"deployOwn":true,"factory":"0x59f30a6703F8Af4FAeAe0e8AFaB14202222F335f"},"compound":{"comptroller":"0x0000000000000000000000000000000000000000","markets":{"0x0000000000000000000000000000000000000000":"0x0000000000000000000000000000000000000000"}},"paraswap":{"deployOwn":true,"contract":"0x22eF7a1De649A8f027384D6a6eDd301AC5FD30D7","authorisedExchanges":{"Kyber":"0x3f906DDA5fa9DA5259d2d865157f5001Ce626e31"}}},"contracts":{"MultiSigWallet":"0xD822e79E2CA49c668f7d97DB1906330bc64459EB","WalletFactory":"0xD7A53fDd46a663C678da0580A66843Ae9b7783Da","ENSResolver":"0xC790c0E4D6811884b52121A5A8d0c0198F515030","ENSManager":"0x75eC71B541b0b2e0f9A4a2bea381c23fd7B0fB0c","TokenPriceProvider":"0xE09C322B745434Fbb2D18c107c13795D4bDF3657","ModuleRegistry":"0x61B2D9C034cD0DAE1738D039cdF9af860ee6B31A","BaseWallet":"0xF346Ee8d0b5738E037DbF22355949411187Da4bC","CompoundRegistry":"0x42ecEbe8946c4954aFc0C2333977adCb243614ed","MakerRegistry":"0x355d26E5f04255F60E910F1aA1251DE5D10996E8","DexRegistry":"0x4A74DE038616791359aEfED19f3f6B75FF6dd340"},"modules":{"GuardianStorage":"0xeE20db41be0CB3D09C5557dDD99d95A4fFf15076","TransferStorage":"0xF8E2BBA1AD8Fd1034DFA1b8Fb6bF5a724fa973ab","GuardianManager":"0x1046E04ED27A22079f69674Bb8bDfbe7c11f43d1","LockManager":"0xCFCE66b4A7904Ad0c4Fe43CB524759C222154BD9","RecoveryManager":"0x8dA7E75A6578F8597eB2c605Fa178F37DbD7E34d","ApprovedTransfer":"0xa6379f76FD8603a44F699EEB42735Ffd00d617D0","TokenExchanger":"0xcF809DD0f9fd4d4b64BcEf64de890BF5579A74A2","NftTransfer":"0x7f7d740Aaf4996C6F4718Fd9003Bb3CC565A11dC","MakerManager":"0x2718cAce241213B18F67C6d66B245EE77DB0Ce2a","TransferManager":"0xc49b23b3403fcEDa2b512D2178fA7020E8f5227E","CompoundManager":"0x55C914A069a49c4163b1dC5FbB216e3ae0339f98","MakerV2Manager":"0x950B78F4B2ecA78C552F4797c0cf10d1aaDDAD2c","LimitStorage":"0xb9dC99d37D77B915768029767C3dC0ff004540E5","TokenPriceStorage":"0x5dA4F81c17776495e3d7999ad9217748D257f074","RelayerModule":"0x171eb25c80da6adF0436668126285ca51e37eB30","RelayerManager":"0x6E5727685D31C3d9bE0c19F5A635551733567e15","VersionManager":"0xf4146041EdE0B35A6573DA0D820cA6Da785dFf17","LockStorage":"0x731a09EEda7ed25080A5f92F117076f6a8ED0E14","TokenPriceRegistry":"0x7e7acD252559062f0E6355847f4Fbb1c62872f14"},"gitCommit":"29a0aee0fd3e30c570e440b69dc14e7ca93755c9"}
